requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pydantic==1.8.2
python-dotenv==0.19.1
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
